+++ packages/dakit_flutter/lib/dakit_flutter.dart
@@ -5,6 +5,7 @@ export 'package:dakit_api/dakit_api.dart';
 export 'package:dakit_core/dakit_core.dart';
 
 export 'src/app_links_callback_source.dart';
+export 'src/client_secure_storage.dart';
 export 'src/dakit_oauth_client.dart';
 export 'src/background_transfer_manager.dart' show BackgroundTransferManager;
 export 'src/secure_pending_authorization_store.dart';